+++ lbrynet/lbrynet_daemon/Apps/LBRYOSXStatusBar.py
