+++ lbrynet/lbrynet_daemon/LBRYDaemon.py
@@ -152,8 +152,8 @@ class LBRYDaemon(xmlrpc.XMLRPC):
         d.addCallback(lambda _: self._setup_lbry_file_opener())
         d.addCallback(lambda _: self._setup_query_handlers())
         d.addCallback(lambda _: self._setup_server())
-        d.addCallback(lambda _: self._update() if self.check_for_updates == "True" and sys.platform == "darwin"
-                                else defer.succeed(None))
+        # d.addCallback(lambda _: self._update() if self.check_for_updates == "True" and sys.platform == "darwin"
+        #                         else defer.succeed(None))
         d.addCallback(lambda _: self._setup_fetcher())
         d.addCallback(lambda _: _disp_startup())
         d.callback(None)
