--- lbrynet/lbrynet_daemon/LBRYURIHandler/setup.py
@@ -1,19 +1,0 @@
-"""
-This is a setup.py script generated by py2applet
-
-Usage:
-    python setup.py py2app
-"""
-
-from setuptools import setup
-
-APP = ['LBRYURIHandler.py']
-DATA_FILES = []
-OPTIONS = {'argv_emulation': True}
-
-setup(
-    app=APP,
-    data_files=DATA_FILES,
-    options={'py2app': OPTIONS},
-    setup_requires=['py2app'],
-)
