+++ setup_uri_handler.py
@@ -1,0 +1,22 @@
+from setuptools import setup
+import os
+
+APP = [os.path.join('lbrynet', 'lbrynet_daemon', 'Apps', 'LBRYURIHandler.py')]
+DATA_FILES = []
+OPTIONS = {'argv_emulation': True,
+           'plist': {
+               'CFBundleURLTypes': [
+                    {
+                    'CFBundleURLTypes': 'LBRYURIHandler',
+                    'CFBundleURLSchemes': ['lbry']
+                    }
+               ]
+           }
+        }
+
+setup(
+    app=APP,
+    data_files=DATA_FILES,
+    options={'py2app': OPTIONS},
+    setup_requires=['py2app'],
+)
